provide zvol2/zvol_thinlv2 storage plugins
The existing zvol plugin mimics the LVM world. There a snapshot also
creates a block device. To have the same behavior in the zfs world, a
snapshot also creates a zfs clone (i.e, the block device). On restore,
we created a second snapshot from that block device and then used that
one as the new backing device for the new resource.

This mimicking of LVM is not necessary. Create one snapshot on
"create-snapshot". On "restore-snapshot" use that one to create the new
backing device.

The old version works, but generates unnecessary snapshots/clones. New
users should use this version, for example for the thinly-provisioned
version:
storage-plugin = drbdmanage.storage.zvol2_thinlv.ZvolThinLv2

These two versions of the zfs plugin are incompatible. If you only used
zfs (without snapshots), you can switch to the new one.