leader: implement different error handling strategies
So far actions are retried until a (configurable) maximum error count is
reached. Then the resource/assignment is ignored until a "resume-all"
gets triggered manually.

This commit introduces a more fine-grained contol over the
error-handling strategy. The strategy is set via "edit-config", where
the following settings are possible:

"error-handling-strategy" is one of:
  "resume-no": the current behavior
  "resume-max-backoff": exponential backoff until a maximum is reached
  "resume-interval": resume with the given interval

The given strategy is also the key to set a value for
"resume-max-backoff" and "resume-interval". Times are given in minutes.

For example:
error-handling-strategy = resume-interval
resume-interval = 30

Issues the equivalent of "resume-all" every 30 minutes if there are
failures.

error-handling-strategy = resume-max-backoff
resume-max-backoff = 65

The first resume is triggered after 2 minutes, the next after 4
additional minutes, the next after 8 minutes and so on. If the next
target time is larger than "resume-max-backoff" (in the given example
when it is set to 128), it is reset to 2.

The strategy is only executed if there are failures drbdmanage detected,
otherwise all counters are reset to 0.