add more function to wrappers
These wrappers are used to ping/pong between server and cli, especially
to tell the client that the server is currently not ready.

When drbdmanaged was shut down, and it was started with "edit-config",
it ran into a dbus error because the signature was wrong because the
entry in the wrappers was missing.

Fix up the ones used in "edit-config".