client: print "Waiting for" messages to stderr
In the old times dbus commands timed out frequently when the server was
busy (because of startup, or it holds a lock and did not accept new
commands,...). This was then changed in a way that the server can answer
with "busy"-messages and the client then prints a "Waiting for" message
and retries several times.

The OpenNebula driver uses to drbdmanage CLI to execute commands and
sometimes also queries information. In busy clusters, especially if
executed on a satellite node, it happened frequently that the whole
process takes a bit longer and one or two "Waiting for" messages got
printed. Unfortunately these were printed to stdout and therefore got
mixed with the output of "query commands". The result was that actually
successfully deployed resources got removed again.

Print these messages to stderr. It should have been like this from the
beginning.

OpenNebula users (and users that rely on the output of the drbdmanage
CLI) have to upgrade. All other official plugins are not affected as
none of them uses the CLI.