client helper: add retry logic
Due to the improved cluster-wide locking, it happens that nodes,
especially satellites, have to retry actions more frequently. Add that
logic, which we had in the CLI to the client helper class.

This is an important update for users of the client helper class, namely
the Docker volume plugin and OpenStack.