Changed LvmThinLv space calculation
Space is now calculated by subtracting data_percent of the size
of the thin pool from the total size of the thin pool.
meta_percent is ignored.
The thin pool may still run out of meta data space, and in this
case, it will probably be unable to allocate new storage devices
although enough free space is shown in drbdmanage.

It is unclear, whether or not the behavior is identical for all
versions of thin LVM, or even what exactly LVM's behavior is
supposed to be.

This is a best guess attempt at reporting a free space number that
may be reasonably close to the actual free space in the pool.