auto-deploy: fix case where resource ends in all diskless
So far auto_deploy() always used the overall assignment count including
diskless nodes. This could lead to the following situation:

diskful: 2
diskless: 3
deploy 2 --with-clients

Then the assignment count was calculated as 5 and the logic tried to
reduce it. As the resource (with count 5) is "over deployed", DM started
to convert "some nodes" (the two diskful) to diskless nodes and finally
assigned the rest diskless. => all diskless.

Fix that in a non-intrusive way. This showed to work as expected and
especially fixes the case where the OpenNebula driver creates a resource
from a snapshot. Because in this snapshot-restore scenario, the resource
starts in such a mixed diskful/diskless state (intentional), and then
the driver calls an auto_deploy().

Still, a rewrite with diskless nodes in mind would be required to fix
all edge cases. For example a "deploy 2 --with-clients" with a later
"deploy 2" does not remove the diskless nodes. Not a regression as this
never worked. Too much effort (and potential for regressions) for the
basically EoL drbdmanage...